[codemirror] Separate factory from editor
